Hopefully fixed last level
When I did the final test of my game yesterday, everything was going perfect, so I uploaded it and submitted it.
This morning, when I gave it another test just to reflect on how I did, the second level was completely pink and gave a bunch of errors. I has recreated the last level, along with a video to show it (if it still doesn't work for some reason)